requirements.txt.txt
openai==0.27.0
streamlit==0.89.0
youtube-transcript-api==0.4.1
gTTS==2.2.3
python-dotenv==0.19.1
